Refactor to use hierarchical identifiers.
Short, meaningful names are important. The default model for Structurizr uses flat identifier naming, so
we can't reuse the same name anywhere in the model.

For larger models, it's useful to use short names that make sense in the context of that thing's parent. As an example, 'mysql' often unambiguously identifies the persistence store for a single service.

This change enables hierarchical modelling. All identifiers need to be referenced with their system and container parent, e.g.

- publishing_platform // the publishing system
- publishing_platform.content_publisher // the content publisher container
- publishing_platform.content_publisher.mysql // the MySQL database within the content publisher container

Watch out for smells! If you regularly need to use long identifiers it may be a sign there's tight coupling in the system